% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot_outliers.R
\name{crossplot_outliers}
\alias{crossplot_outliers}
\title{To get the outlier data form a ggplot}
\usage{
crossplot_outliers(p, num.outliers = NULL, percentile.outliers = NULL,
  plot.vars = c("x", "y"))
}
\arguments{
\item{p}{a crossplot \code{ggplot2} plot. Can be generated from
\code{crossplot}}

\item{num.outliers}{the number of outliers. Defaults to \code{NULL}}

\item{percentile.outliers}{the percentile used to measure
outliers. Defaults to \code{NULL}}

\item{plot.vars}{the plot variables for which the outliers will be
calculated. These variables are from \code{ggplot2} mappings
and can include \code{x}, \code{y}, \code{size}. The default is
\code{c("x", "y")}}
}
\value{
a \code{data.frame} with the plot data for the outliers
}
\description{
Based of plot variables from the user, \code{crossplot_outliers}
returns the plot data corresponding to the outliers. The user can
imput either imput the number of desired outliers,
\code{num.outliers}, or the percentile corresponding to the
outliers \code{percentile.outliers}, but not both
}
\examples{
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point(aes(size = wt))
##Get the outliers -- defaults to outliers for just "x" and "y"
crossplot_outliers(p, num.outliers = 2)
crossplot_outliers(p, percentile.outliers = .05)
##Specifying the plot variables
crossplot_outliers(p, num.outliers = 2, plot.vars = c("x", "y", "size"))
crossplot_outliers(p, percentile.outliers = .05, plot.vars = c("x", "y", "size"))
}

