% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplotr_main.R
\docType{package}
\name{crossplotr-package}
\alias{crossplotr}
\alias{crossplotr-package}
\title{Create easy cross sectional plots with ggplot2 and cowplot}
\description{
crosspolotr allows for easy creatation of cross-sectional plots
}

