% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_no_na_values}
\alias{data_no_na_values}
\title{Check that the data does not have and NA values}
\usage{
data_no_na_values(data, env = parent.frame())
}
\arguments{
\item{data}{the data set}

\item{env}{the environment. Defaults to \code{parent.frame}}
}
\value{
a data frame with no missing values for the variables of
    interest
}
\description{
Return a dataframe with no NA values for the variables of interest
}
