% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reg_mean}
\alias{reg_mean}
\title{Mean of a variable using the \code{lm} function}
\usage{
reg_mean(df, var, weights.var = NULL)
}
\arguments{
\item{df}{\code{data.frame} of data}

\item{var}{a string with the variable name}

\item{weights.var}{a string with variable to be used as
weights. Defaults to \code{NULL}}
}
\value{
an object from \code{lmtest::coeftest} with the output from
    the regression in a 1x4 matrix including the mean (the
    estimate), the standard error, the t-value, and the p-value
}
\description{
return is the output from \code{lmtest::coeftest} which includes
the mean, the standard error, the t-stat and p-value
}
