% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot_print_stats.R
\name{crossplot_print_stats}
\alias{crossplot_print_stats}
\title{A convience function to print the plot stats and a regression line}
\usage{
crossplot_print_stats(p, stats = c("slope", "r.squared"), text.pos,
  sprintf.format = "\%.2f", xlabel = NULL, ylabel = NULL,
  log.reg = FALSE, weighted = FALSE, reg.label.se = TRUE,
  reg.color = "black", reg.linetype = "solid")
}
\arguments{
\item{p}{a crossplot \code{ggplot2} plot. Can be generated from
\code{crossplot}}

\item{stats}{character vector with the stats that will be printed
(in order). Possible statistics include \code{"slope"},
\code{"intercept"}, \code{"r.squared"}, \code{"adj.r.squared"},
\code{"mean.x"}, \code{"mean.y"} \code{"var.x"},
\code{"var.y"}, \code{"sd.x"}, \code{"sd.y"}. Defualt is
\code{c("slope", "r.squared")}}

\item{text.pos}{the text position to be passed to
\code{cowplot::draw_label()}}

\item{sprintf.format}{character string with the \code{sprintf}
format}

\item{xlabel}{the label for the x-axis variable}

\item{ylabel}{the label for the y-axis variable}

\item{log.reg}{logical. If set to \code{TRUE}, the regression will
be estimated in logs}

\item{weighted}{logical. If set to \code{TRUE}, the regression and
all statistics will be weighted using the \code{size} variable
form the \code{ggplot2} plot}

\item{reg.label.se}{logical. If set to \code{TRUE}, the regression
standard errors will be printed}

\item{reg.color}{a string with the \code{ggplot2} color for the
regression line}

\item{reg.linetype}{a string with the \code{ggplot2} linetype for
the regression line}
}
\value{
a \code{ggplot2} plot with the regression line and label
    added
}
\description{
A convience function to print the plot stats and a regression line
}

